+++ src/Solver/MatrixStorageFormats/CoordinateFormat.c
@@ -31,7 +31,7 @@ CoordinateFormat_t* (CoordinateFormat_Create)(Mesh_t* mesh,Options_t* options,co
       int nnz = CoordinateFormat_GetNbOfNonZeroValues(cfmt) ;
       /* The length required by ma38 must not be lower than 2*nnz */
       double ff = Options_GetFillFactor(options) ;
-      int lv  = floor(ff*(2*nnz)) ;
+      int lv  = (int) floor(ff*(2*nnz)) ;
       double* v = (double*) Mry_New(double[lv]) ;
       
       CoordinateFormat_GetLengthOfArrayValue(cfmt) = lv ;
@@ -46,7 +46,7 @@ CoordinateFormat_t* (CoordinateFormat_Create)(Mesh_t* mesh,Options_t* options,co
       int n = Mesh_GetNbOfMatrixColumns(mesh)[imatrix] ;
       /* The length required by ma38 must not be lower than 3*nnz+2*n+1 */
       double ff = Options_GetFillFactor(options) ;
-      int lindex = floor(ff*(3*nnz + 2*n + 1)) ;
+      int lindex = (int) floor(ff*(3*nnz + 2*n + 1)) ;
       int* index = (int*) Mry_New(int[lindex]) ;
       
       CoordinateFormat_GetLengthOfArrayIndex(cfmt) = lindex ;
